requirements.txt
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
clickclick==1.2.2
colour==0.1.5
connexion==2.3.0
Flask==1.0.3
Flask-SQLAlchemy==2.3.0
GeoAlchemy2==0.6.3
hashids==1.2.0
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
MarkupSafe==1.1.1
Naked==0.1.31
openapi-spec-validator==0.2.7
pycrypto==2.6.1
pyrsistent==0.15.2
PyYAML==5.1.1
requests==2.22.0
shellescape==3.4.1
six==1.12.0
SQLAlchemy==1.3.5
SQLAlchemy-Utils==0.34.0
typing==3.7.4
urllib3==1.25.3
Werkzeug==0.15.4
